requirements.txt
boto3
botocore
tqdm
requests
Pillow
pillow-heif
jupyterlab
moviepy==1.0.3
xmltodict
pyyaml